requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-rest-swagger==2.2.0
djangorestframework==3.12.2
idna==2.10
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-decouple==3.4
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
